feat: add sprites for race condition slime
