+++ src/main.cpp
@@ -1,5 +1,5 @@
 #include "entities/player/player.h"
-#include "entities/race_condition_enemy.h"
+#include "entities/race_condition_slime.h"
 #include "world/world.h"
 #include <SFML/Graphics.hpp>
 #include <cmath>
@@ -20,8 +20,8 @@ int main()
 
 	Player player;
 	World world;
-	RaceConditionEnemy race_condition_enemy_1({5 * 32.f, 15 * 32.f});
-	RaceConditionEnemy race_condition_enemy_2({11 * 32.f, 15 * 32.f});
+	RaceConditionSlime race_condition_enemy_1({5 * 32.f, 15 * 32.f});
+	RaceConditionSlime race_condition_enemy_2({11 * 32.f, 15 * 32.f});
 
 	// world.loadFromJson("data/maps/test.json");
 	world.loadFromTMJ("data/maps/test.tmj");
